With (generic) functional options.
We usually define functional options as methods on a singleton, but
methods only support generic receivers, not generic arguments, so these
options are more traditional package-level functions. We really don't
need generics for this. The next commit will probably undo the generics.

IMHO, functional options in this case are kinda silly since the three
provided are mutually exclusive of one another so this all feels like
a mute point.